Add files via upload
This script logs in to the amazon account and searches for snow boots and select a boot and the size required and adds to the cart. After that it adds a dress shirts, select and adds the size and then its added to the cart. Then logs out of the amazon account.